% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-health_departments.R
\docType{data}
\name{health_departments}
\alias{health_departments}
\title{General overview of German Health Authorities}
\format{
A tibble with 399 rows and 8 columns
\describe{
\item{authority_id}{Technical identifyer of the health authority, imporant for merging with other data sets}
\item{name}{Name of the health authority}
\item{department}{Type of health authority}
\item{street}{Street name}
\item{postalcode}{zip code}
\item{place}{City}
\item{long}{longitude of the health authority}
\item{lat}{latitude of the health authority}
}
}
\usage{
health_departments
}
\description{
This data is generated from RKI PLZ Tool
\url{https://www.rki.de/DE/Content/Infekt/IfSG/Software/software_node.html}
geolocations are generated using the package \code{ggmap}.
}
\examples{
health_departments

}
\keyword{datasets}
